Restructured the code in multiple file. Added some comments. Changed the layout of some frames. Updated the graphic.
