+++ thermocepstrum_gui/core/control_unit.py
@@ -1,6 +1,6 @@
 # -*- coding: future_fstrings -*-
 
-'''
+"""
 --------------------------------------------
     Thermocepstrum graphic user interface
 
@@ -9,7 +9,7 @@
 
 This file contains the functions that the GUI will directly call to start new
 multithread operations and to make requests to the thermocepstrum calculus unit.
-'''
+"""
 
 import os
 from . import settings
@@ -18,8 +18,23 @@ import thermocepstrum as tc
 import numpy as np
 from thermocepstrum_gui.utils import Graph
 
+try:
+    from thermocepstrum.utils.utils import PrintMethod
+except ImportError:
+    from thermocepstrum_gui.utils.utils import PrintMethod
+
+# Init print method
+log = PrintMethod()
+
+
+# -------- DATA SECTION --------
 
 class Data:
+    """
+    This class is used as a structure to
+    contains all the variables that the software use.
+    """
+
     CURRENT_FILE = ''
     loaded = False
 
@@ -44,15 +59,53 @@ class Data:
     psd_filter_width = 0.1
 
 
+# -------- GRAPH SECTION --------
+
+
+"""
+This section contains the functions that deal with
+the graph manager.
+"""
+
+
+# Setup graph manager
 gm = Graph.GraphManager()
 
 
 def set_graph(axis_, func, **kwargs):
+    """
+    This function request to the graph manager to generate
+    a new graph to be displayed in the axis_ screen.
+
+    :param axis_: the location to draw the new graph.
+    :param func: the function to generate the graph.
+    :param kwargs: other arguments to pass to the file manager.
+    :return axis: the new graph.
+    """
+
     axis = func(axis=axis_, external_object=Data, **kwargs)
     return axis
 
 
+# -------- FILES AND LOAD SECTION --------
+
+
+"""
+In this section there are the functions that are used to load the data
+that the software will use and some functions to work with files.
+
+The starter settings are loaded from the .ini file that contains the
+values that need to be stored.
+"""
+
+
 def get_file_size(path):
+    """
+    This function calculate the file size.
+
+    :param path: the location of the file.
+    :return: the file size.
+    """
     file_size = os.path.getsize(path)
     if file_size >= 1000000:
         file_size //= 1000000
@@ -62,81 +115,21 @@ def get_file_size(path):
         return f"{file_size} MB"
 
 
-def secure_exit(main_window):
-    '''
-    This function allow a secure exit from the execution
-    of the software.
-    '''
-
-    # todo: Add multithread processes safe exit and a log file
-    for process in main_window.open_windows:
-        process.destroy()
-
-    exit()
-
-
-def get_cor_index(arr1, arr2, corr_key):
-    i = arr2.index(corr_key)
-    return arr1[i], i
-
-
-def get_temp(jdata, selected_key):
-    if selected_key in jdata:
-        temperature = np.mean(jdata[selected_key])
-        temperature_std = np.std(jdata[selected_key])   # this is wrong (needs block average)
-        print(' Mean Temperature (computed):  {} K  +/-  {}\n'.format(temperature, temperature_std))
-    # elif 'Temp_ave' in jdata:
-    #     temperature = jdata['Temp_ave']
-    #     if 'Temp_std' in jdata:
-    #         temperature_std = jdata['Temp_std']
-    #         print(' Mean Temperature (file):      {} K  +/-  {}'.format(temperature, temperature_std))
-    #         # logfile.write(' Mean Temperature (file):      {} K  +/-  {}\n'.format(temperature, temperature_std))
-    #     else:
-    #         print(' Mean Temperature (file):      {} K'.format(temperature))
-    #         # logfile.write(' Mean Temperature (file):      {} K\n'.format(temperature))
-    else:
-        temperature = -1
-        # raise RuntimeError('No Temp key found. Please provide Temperature (-T).')
-
-    return temperature
-
-
-def get_volume(jdata, structurefile):
-    if structurefile is not None:
-        _, volume = tc.i_o.read_lammps_datafile.get_box(structurefile)
-        print(' Volume (structure file):    {} A^3'.format(volume))
-        # logfile.write(' Volume (structure file):    {} A^3'.format(volume))
-    elif 'Volume' in jdata:
-        volume = jdata['Volume']
-        print(' Volume (file):    {} A^3'.format(volume))
-        # logfile.write(' Volume (file):    {} A^3\n'.format(volume))
-    else:
-        volume = -1
-        # raise RuntimeError('No Volume key found. Please provide Volume (-V) of structure file (--structure).')
-
-    return volume
-
-# -------- LOAD SECTION --------
-
-'''
-In this section there are the functions that are used to load the data
-that the software will use.
-
-The starter settings are loaded from the .ini file that contains the
-values that need to be stored.
-'''
-
-
 def load_path():
-    '''
-    This function loads the paths where the files are stored
-    '''
+    """
+    This function loads from the .ini file the paths
+    where the file will be stored.
+    """
+
+    # Verify that the file exists otherwise generate it
     if not os.path.exists('thcp.ini'):
         set_defaults()
 
+    # Read the file
     with open('thcp.ini', 'r') as file:
         settings_file = file.read().split('\n')
 
+        # Load the paths
         if settings_file[0]:
             for setting in settings_file:
                 if setting:
@@ -158,14 +151,16 @@ def load_path():
 
 
 def set_defaults():
-    '''
+    """
     This function is used to set the default values of the .ini file
-    '''
+    """
+    # Create the file
     with open('thcp.ini', 'w') as file:
         defaults = (('DP', 'data'),
                     ('OP', 'outputs'),
                     ('LP', 'logs'))
 
+        # Write the defaults paths
         for setting in defaults:
             try:
                 os.mkdir(os.path.join(settings.BASE_PATH, setting[1]))
@@ -177,11 +172,20 @@ def set_defaults():
 
 
 def load_keys(inputfile):
+    """
+    This function is used to load the header keys of the selected file.
+
+    :param inputfile: the path of the selected file.
+    :return:
+    """
+
     jfile = tc.i_o.TableFile(inputfile, group_vectors=True)
     return jfile.all_ckeys
 
 
-def load_data(inputfile,input_format,selected_keys,temperature=None,NSTEPS=0,START_STEP=0,run_keyword='',units=None,DT_FS=None,volume=None,psd_filter_w=None,axis_=None, structurefile=None, descriptions=[]):
+def load_data(inputfile, input_format, selected_keys, temperature=None, NSTEPS=0, START_STEP=0,
+              run_keyword='', units=None, DT_FS=None, volume=None, psd_filter_w=None, axis_=None,
+              structurefile=None, descriptions=[]):
 
     Data.temperature = temperature
     Data.volume = volume
@@ -247,7 +251,77 @@ def load_data(inputfile,input_format,selected_keys,temperature=None,NSTEPS=0,STA
         return -1, emsgs
 
 
+# -------- UTILITY SECTION --------
+
+
+"""
+This section contains some utility functions.
+"""
+
+
+def secure_exit(main_window):
+    '''
+    This function allow a secure exit from the execution
+    of the software.
+    '''
+
+    # todo: Add multithread processes safe exit and a log file
+    for process in main_window.open_windows:
+        process.destroy()
+
+    exit()
+
+
+def get_cor_index(arr1, arr2, corr_key):
+    i = arr2.index(corr_key)
+    return arr1[i], i
+
+
+def get_temp(jdata, selected_key):
+    if selected_key in jdata:
+        temperature = np.mean(jdata[selected_key])
+        temperature_std = np.std(jdata[selected_key])   # this is wrong (needs block average)
+        print(' Mean Temperature (computed):  {} K  +/-  {}\n'.format(temperature, temperature_std))
+    # elif 'Temp_ave' in jdata:
+    #     temperature = jdata['Temp_ave']
+    #     if 'Temp_std' in jdata:
+    #         temperature_std = jdata['Temp_std']
+    #         print(' Mean Temperature (file):      {} K  +/-  {}'.format(temperature, temperature_std))
+    #         # logfile.write(' Mean Temperature (file):      {} K  +/-  {}\n'.format(temperature, temperature_std))
+    #     else:
+    #         print(' Mean Temperature (file):      {} K'.format(temperature))
+    #         # logfile.write(' Mean Temperature (file):      {} K\n'.format(temperature))
+    else:
+        temperature = -1
+        # raise RuntimeError('No Temp key found. Please provide Temperature (-T).')
+
+    return temperature
+
+
+def get_volume(jdata, structurefile):
+    if structurefile is not None:
+        _, volume = tc.i_o.read_lammps_datafile.get_box(structurefile)
+        print(' Volume (structure file):    {} A^3'.format(volume))
+        # logfile.write(' Volume (structure file):    {} A^3'.format(volume))
+    elif 'Volume' in jdata:
+        volume = jdata['Volume']
+        print(' Volume (file):    {} A^3'.format(volume))
+        # logfile.write(' Volume (file):    {} A^3\n'.format(volume))
+    else:
+        volume = -1
+        # raise RuntimeError('No Volume key found. Please provide Volume (-V) of structure file (--structure).')
+
+    return volume
+
+
 def update_info(frame):
+    """
+    This function is used to write the info in
+    the info screen.
+
+    :param frame: the info screen.
+    """
+
     frame.clear()
     frame.write('file name:        {}'.format(Data.jfile.filename))
     frame.write('file size:        {}'.format(get_file_size(Data.jfile.filename)))
